Add a separate service to serve www.snapcraft.io
and redirect.

We initially wanted to use [the redirect-from-to-www annotation](https://github.com/kubernetes/ingress-nginx/blob/master/docs/user-guide/annotations.md#redirect-from-to-www), as this is nice and neat, but unfortunately that [isn't able to serve a TLS certificate for the redirected domain](https://github.com/kubernetes/ingress-nginx/issues/605).

Until that's fixed, we have to use this solution, [as others are doing](https://github.com/kubernetes/ingress-nginx/issues/605#issuecomment-334931436).

(It would also help to make this more DRY if we were able to [specify multiple hosts in a http rule](https://github.com/kubernetes/kubernetes/issues/43633))