Set nginx redirect code from 308 to 301
In the latest version of nginx-ingress-controller, it sets the redirect
code to 308. Is causes some issues with social media bots/crawlers which
don't seem to recognise the 308 HTTP code.

This will set the redirect code to 301 but it is not the preferrend
situation in the long term. As soon as it is reasonable to revert this
change, we should do so.
The docs for this setting can be found at:
https://github.com/kubernetes/ingress-nginx/blob/master/docs/user-guide/configmap.md#http-redirect-code