Add "tag" option, convert arguments to options
Previously it's not been outwardly clear that the QA script will attempt
to release an image that it expects to find at
prod-comms.docker-registry.canonical.com/{project}:latest.

There is also the problem that usually in the basic flow of initially
creating an image for a new service, we don't necessarily publish to the
"latest" tag. It's more common to publish to a tag like
"3018b561625ad213f67715d7121b15c16056e8d1-1519153583".

Therefore I think it's much better to be declarative about requiring the
user of `qa-deploy` to provide an image tag. This makes it much more
clear that that tag has to exist, and for the person that ends up
copying and pasting the "./qa-deploy --production example.com --tag
3018b561625a", it's more clear what's going on if they feel like
debugging any problems.