Add all canonical domains to NO_PROXY
For each domain we use the `{domain}.{tld}` syntax *as well as* the
`.{domain}.{tld}` syntax. This is because the format of NO_PROXY is not
standardised, and I can find no confirmation that `.example.com` would
normally match `example.com` itself. Therefore, to be in the safe side,
I am explicitly including each main domain.

https://superuser.com/questions/1300625/best-way-to-specify-domain-and-all-sub-domains-in-no-proxy