Reset the imagePullPolicy to use the default
We initially set this pull policy to "Always" so that we could do "kubectl set image xxxx" with the *same* commit ID as the image tag and still trigger a roll-out.

However, now we add timestamps to the image names, meaning we can trigger rollouts whenever we need just by pushing an image with a new timestamp, so this is no long an issue.

We've now had a few outages when Kubernetes can't communicate with the registry. "imagePullPolicy: Always" would certainly exacerbate these problems, because it prevents kubernetes from using a local image if one exists.

Since I don't think we need imagePullPolicy: Always any more, I'm removing all "imagePullPolicy: Always" configuration, so it will use the default of "IfNotPresent".