Configs for 360.canonical.com
This is a little special for a few reasons:

- The normal meta name we give services etc. would in this case be
`360-canonical-com`, but this is not a valid name in this case 'cos it
starts with a 3. So instead we use `threesixty-canonical-com` in all
cases where we can't use the actual domain name.
- It needs a database, so it looks for a `threesixty-database-url`
secret for the location of the database
- It needs access to some CSV files for employee data to initially
provision the database, so it looks for a `threesixty-employee-data`
secret
- Liveness and readiness checks use `/_status/ping` instead of
`/_status/check` because at the moment `/_status/check` redirects you to
SSO