Add configs for mongoose.ubuntu.com
This will be the domain under which we test the new beta changes for
www.ubuntu.com. It's called Mongoose, because we need to use a unique,
sacrificial domain name so that domain can then simply be redirected back to www.ubuntu.com
after the test completes.

The mongoose ingresses are in their own file, as this is the only way to
give them their own HTTPS certificate (like what we did for
insights.ubuntu.com).

I've also added `X-Robots-Tag: noindex` to mongoose.ubuntu.com and
beta.ubuntu.com, as we don't want search engines sending people to
either.