Disable "includeSubdomains" in HSTS
This disables the "includeSubDomains" setting in HSTS for all domains,
even though the only one where we really *need* to turn it off is
actually canonical.com.

However, after much searching, there doesn't appear to be any way
to turn off "includeSubDomains" on a per-ingress level.

If we really wanted all other sites to have "includeSubDomains", but
turn it off only in canonical.com, the only way to achieve it would be
to turn HSTS off globally (`hsts: "false"`), and then manually add
the `strict-transport-security` to every ingress via config snippets.

This seems more risky, as we could easily forget to enable it for a
domain, and certainly more cumbersome. On balance, the
"includeSubDomains" setting isn't very important for most domains,
and arguably dangerous for some, so in the circumstances
it's probably best to simply disable it globally.

If Kubernetes add the ability to trivially tweak the setting at the
ingress level, we may change this global setting in the future.